roof_survival_detection/requirements.txt
ultralytics
opencv-python
numpy
torch
torchvision
transformers
accelerate
python-telegram-bot
pillow
requests
scipy

